Use Provider<PeerInfo> instead of Configuration.getUrl()
By using Provider<PeerInfo> we make it possible to provide the URL of
the other peer node in a different way than reading it from the plugin's
global configuration.

To keep compatibility with the existing configuration of the secondary
peer node URL, this change implements Provider<PeerInfo> in
PluginConfigPeerInfoProvider which reads the secondary peer's URL from
the plugin's configuration file.

The assumption that there is only one secondary peer node still stays.
Introducing the concept of N peer nodes (N > 2) may be done in a future
change.

Change-Id: I2df5c16c951023df01a6b03ddb53776d52c607d7