Add file based websession sharing
The implementation is a fork of the websession-flatfile plugin. This
fork was required to provide this all-in-one plugin allowing to run an
active/passive setup and keep the websession-flatfile plugin as is, i.e.
not deprecating it.

This implementation is similar to the original but reworked to be more
testable and many unit test cases added to keep the code-coverage of the
high-availability above 89%.

Change-Id: I302f67e9f3546cdcb8634ec0dd678be73f87f94d