Add support for forwarding cache eviction for custom caches
Allow to specify cache.pattern in the plugin configuration, which
will add additional patterns to the regex to test if a cache eviction
should be forwarded. This will allow caches created by other plugins
to be handled in addition to the default core caches.

For example:

  [cache]
    synchronize = true
    pattern = ^my_cache.*
    pattern = other_cache

Note that evictions for core caches are always forwarded. Specifying
cache.pattern only adds extra matches; it doesn't override forwarding
of eviction for the core caches.

Change-Id: Ia415d53a3c08d744324e88a6f7115a761f94c1f6