Improve the debug log when RestForwarder fails
In case RestForwarder is not able to forward the request
because of exception that is recoverable, in debug logs
there is no info on what is the cause of the problem.

This fix adds the stack trace to debug log in this case.

Change-Id: Icfe5abfe65ff51ff029e0f725c0bb67b349a778a
(cherry picked from commit 1fb2b28743715dd8b4acdbcdc38f7dccee30ed0c)