ConfigurationTest#testGetDefaultSharedDirectory: Don't override mocks
Use the mocked PluginConfigFactory and Config instances defined at class
scope and initialized in the setUp method, rather than defining method
scope instances which override the class scope instances.

Using local instances has no side-effect on stable-2.13, but when this
change is applied on stable-2.14 the test fails due to the fact that
there are other settings mocked in the class scope instances that are
not in the local instance.

Change-Id: Ifb54d7096c51a699531cd954ee8302a31fb8b095