Remove usage of Files.exists and Files.isRegularFile
This is flagged as an issue by Sonarqube: the Files.exists method has
noticeably poor performance in JDK 8, and can slow an application
significantly when used to check files that don't actually exist.

The same goes for Files.notExists, Files.isDirectory and
Files.isRegularFile.

See
https://bugs.openjdk.java.net/browse/JDK-8153414
https://bugs.openjdk.java.net/browse/JDK-8154077

These are not used at runtime by the plugin code, only during the init
so it's not critical but still remove them to reduce the chance of
having this creeping into the plugin code and to make Sonaqube happy.

Change-Id: Ic5fe5f06065878d7c8d680a99ea293c9e0dcd391