Do not refer to KeyUtils internals in tests
KeyUtil is an encoder used in GwtOrm ids, an internal for parsing
strings into IDs. However, high-availability plugin tests should not
be dependent on Gerrit internal implementations if they are not
relevant.

KeyUtil is no longer needed in the internal implementation of the
account indexing and not even in the functional tests.

Change-Id: Ib10d7545a9b1280a0edc7112aad0bb613523b794