Use evicted object key toString
Ibfb40030e was trying to avoid having to setup the encoder which is
required by toString method of some evicted key such as
AccountGroup.UUID.

The problem is it introduced some code which break the encapsulation of
the object by reimplementing differently the toString method. All this
for the sake of removing unit test setup. That reason is not a good
reason to introduce runtime code harder to understand that simply
calling object.toString.

Change-Id: I6c234aa6f0ffaee7c36889cfda04164dc9849134