Use consistent name for Logger instances
In the current code base we use either "log" or "logger" for the name of
the Logger instance. Using a consistent name would make it a bit easier
to grep the code for places where logging is done.

The name "log" is used more commonly than "logger":

  $ git grep "final Logger logger " | wc -l
  6
  $ git grep "final Logger log " | wc -l
  8

so replace instances of "logger" with "log".

Change-Id: I4e6f8a5b685113bf5c26e6ad3d86954c258dc2ce