Fix GroupIndexForwardingIT
The setup was causing the group to be indexed but not the doAction. The
test was passing because wiremock is already up and running before setup
is executed so it was getting a post on the group index endpoint.

A follow up change will rework all the IT test to load the plugin only
once the setup is completed so wiremock only receive request done from
the test method.

Change-Id: I5cb5b1b8d33bfd8d60b36b9cff89513a9ede1da0