Make Wiremock return 204 for any request
Wiremock was only setup to return 204 for the expected request of the
test. This is causing slowdown in the test execution because the test
is generating more forwarding events that the ones expected in the test
and these will retry until retry count or test is completed. These
retries are giving more work than needed to Wiremock thus slowing down
the processing of the expected request.

With this change, the gain is negligible on stable-2.14 but on 2.15, the
overall tests execution is going from few minutes down to 20-30 seconds
as per stable-2.14.

Now that we no longer have unexpected requests, the WireMockRule option
set to false to no fail test on unexpected request can be removed.

Change-Id: If59173437a3db8a9f6e24412b35dc0f9a4df89f0