ForwardedIndexingHandler: Make doDelete abstract
doDelete has a default implementation that throws a "cannot delete"
exception, and it is only overridden in the ForwardedIndexChangeHandler
which actually supports deletion from the index.

However, this default implementation results in two warnings from
Eclipse:

- The id parameter is not used
- The declared IOException is not thrown

Make the doDelete method abstract, and move the exception throwing in
to concrete methods in the derived classes, throwing exceptions with
more specific messages.

Change-Id: Idefa433446f1956ae2e0e275d0996f02c29134df