Forward stream events to message broker
Introduce an event type and allow to publish to different
Kafka topics for different types.

Two events types are defined:
- INDEX_EVENT
- STREAM_EVENT

Adapt the plugin's configuration to the new settings
and update the overview of the plugin and how it is supposed
to be configured and used.

NOTE: Consuming stream events is not supported yet.

Next step will be the support for cache events.

Original-Author: David Ostrovsky <david.ostrovsky@gmail.com>
Change-Id: I3f15623a9666604ed5e79dd84c5072678a961c22