Fix serialization of comment events
This change fixes a bug whereby CommentAddedEvent (as well as other
events having members of type Supplier<?>) were not serialized
correctly, thus causing com.google.gson.JsonSyntaxException at
consumption time.

This change also introduces logging of exception stacktrace to ease
understanding of failures.

As well as testing CommentAddedEvent serialization, this change also
remove the usage of a custom GsonProvider in KafkaEventDeserializerTest
and replaces it with the GsonProvider singleton in order to test the
_real_ runtime serialization rather than a testing-specific, custom one.

Change-Id: I56b7e8d58241c2a6571453dea1cf9e7cbbb9148c