Allow configuration of kafka specific props
Previous this change, all configuration entries were passed to kafka
publisher and consumer even though meaningless and irrelevant.
This was causing warnings such as:

The configuration 'project.list.event.enabled' was supplied but isn't a known config.

This change introduces the ability to configure kafka specific
properties in multi-site.config so that those, and only those will
contribute to configure kafka publisher and subscriber.

Change-Id: Ib53a0b2055f3432a23f9834b77779ec40ee84263