ForwardedIndexProjectHandler: don't use Project.NameKey
Project.NameKey is one of those classes that derives from GWT
because it was used in the old Gerrit UI. It is cursed by
the ORM key serialization internals and may trigger NPE
during tests.

Reduce the creation of a Project.NameKey at the very last moment
so that we can skip all the quirks in the logs.

Change-Id: I42d20709e44449dd530c844f7379a16b17673fc3