ForwardedIndexGroupHandler: do not use UUID as key
The AccountGroup.UUID is based on the GWT-ORM hierarchy
and uses static methods for its printing.

The GWT-ORM stack is not used anymore and there is no reason
why we should be relying on it in our code. The Gerrit API
still rely on the UUID class for reindeixng, so materialise
the strings at the very last moment.

This fixes random NPEs in tests.

Change-Id: I64edc597940f25597cb18488165eaedfe90bbf12