Update patch set 2
Patch Set 2: Code-Review-1

(10 comments)

This looks good, however we should follow the previous pattern with the high-availability plugin:
- Activity => on the multi-site_log (was in httpd_log in high-availability)
- Debug/Errors/Warnings => on the error_log

If the multi-site is healthy, error_log shouldn't have any recurring message about it, apart from start/stop.
If there are problems, the Gerrit admin expects them in the error_log, not in the multi-site_log
If you are debugging, you expect the messages on the console, not in the multi-site_log
If you want to check the message activity, you would process the multi-site_log and not the error_log

Possibly would be best to call it "message_log"?

Patch-set: 2
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1