Add embrional support for split brain detection
Adds a RefOperationValidationListener to intercept and block operations
if they might cause a split brain, i.e. if the oldId of the refs
they are based on is not the currently most recent in the cluster.
The current code doesn't provide a real implementation of the
centralised DB but just an in memory implementation and a no-op one.
The in memory implementation is probably useless but shows the mechanics
of the centralised one.

The Integration Tests are currently just validating that the new code doesn't
introduce any regression.

Feature: Issue 10554
Change-Id: I8351ddb386a9528721c1c9670e3304b40f88a67c