Check for NoteDb migration when loading/testing multi-site
The multi-site plugin is designed to work in conjunction with changes
meta-data on NoteDb under the '/meta' ref. Make sure that Gerrit
changes are migrated when starting the plugin and before running
any integration test.

Relax also the requirement about the number of events received
for asserting only the ones that are relevant to the tests.
NoteDb may generate extra events that are not necessarily interesting
for the multi-site plugin and can be thus safely ignored.

Change-Id: Ia7b9db125e96fafe4eaba1518ce8007ae632d8b2