Adds migration operation mode to allow the ZkSharedRefDatabase to support missing entries
If configured to be in migraiton mode, the ZkSharedRefDatabase will handle failures in updating refs
caused by a missing entry in Zookeeper as the expected consequence of the Zookeeper storage not being
aware of all the objects already in the Git repo but not stored in it.

This is an approach alternative of having a batch process required to initialise the Zookeeper
storage with all existing objects in a given Git repo.

Feature: Issue 10554
Change-Id: I26ae9ef76e288cfe4fa86b65a33b966ce43006a6