Implement with the CommitValidationListener
This version uses the CommitValidationListener to check
check that data is stored in ZK with changes.

Currently the RefOperationValidationListener does not provide us
a way to tap into all the events that trigger an alteration to
the git repository. A good example of this are the change references
'refs/changes' that get commited into git but never trigger the
RefOperationValidationListener.

The CommitValidationListener on the other hand provides a way to
tap into most events that trigger a modification of the git repository.
A shortcoming of this implementation is that when a change is created
the _refName_ within a CommitReceivedEvent has the value
'refs/heads/master' rather than the expected 'refs/changes/'.

This implementation is an alternative to the proposed change
of triggering RefOperationValidation for 'refs/changes'.
(https://gerrit-review.googlesource.com/c/gerrit/+/218581)

Change-Id: Ia124b92c07fdd660ea7495c0612aa45fbd6320d3