Remove the thread-local event forwarding logic
Event publish and subscribe happen in different threads and thus the
old logic of re-binding the even forwarder and check for the local
thread storage isn't used anymore.

This is a leftover from the high-availability plugin, because
the protection for non-forwarding from local events is done based
on gerrit server's instance-id and not thread local in multi-site
plugin.

Change-Id: I0a448403c89275fde006e9294e5cfc6a9693280a