Turn the multi-site plugin into a libModule
Abandon the plugin packaging style and turn the whole project into
a libModule. The @PluginName annotation isn't available anymore
and thus needs to be replaced with a static string.

The repackaging is needed in preparation for the injection of an
alternative Git repository manager to intercept and wrap all
ref-updates and check them against the shared ref-db.

Re-binding the Git repository manager is done very deeply in the Gerrit
daemon startup process and does not allow any modification through
plugins.

The introduction of a libModule allows influencing the Guice bindings
at the core of Gerrit daemon startup.

Change-Id: I302e2c303ca145f951426912a71e87b44082c85b