Bind GitRepositoryManager and associated factories
Intercept ref-updates at GitRepositoryManager level, wrapping the
repository, ref-database, ref-update and batch ref-update.

The InSyncChangeValidator is not useful anymore because it was covering
a very limited set of use-cases.

NOTE: The integration test does not cover yet the GitRepositoryManager
binding for the refs validation: this is due to the lack of support
of libModules inside the current acceptance test framework in Gerrit.

Change-Id: I4ecdfb1d4a779ab54a15831a692e4b2b697a3021