Update patch set 2
Patch Set 2:

Now I recall why I have removed all this logic: it just did not bind correctly when using multi-site as a libModule.

I naively thought we did not need that anymore because we have the instanceId. However, that instanceId hasn't been introduced in Gerrit core and thus the whole of Gerrit is unaware of it and keeps on looping back the events to the same instance.

This rever alone *won't work* and I'll follow-up with further fixes to make it work for libModule. I won't amend this clean-cherry-pick though so that we can track what has been fixed to make it work.

Patch-set: 2
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>