Differentiate the multi-site Gson from Gerrit's
Gerrit uses an explicit binding of the Gson serializer to be used
internally, not to be confused with the Gson serializer used for
sending and receiving events to the Broker.

Introduce a new specific annotation BrokerGson and use it throughout
the code.

Eliminate also the indirection through a Provider<Gson> which is an
unnecessary complication.

Change-Id: I387d50f70e2c5ba9dc229a95aa5b93497be0a799