Validate replication.config at Gerrit startup
When using a multi-site setup, the gerrit.replicateOnStartUp flag
on the replication.config will create a split-brain across the
cluster because assumes that the off-line master node is the
source of truth for the world, which is clearly NOT the case.

In a multi-site setup, there are always nodes up-and-running
somewhere. If a node was off, it will be very likely to be behind
and should *never* force the other nodes to comply with its
offline status.

Change-Id: I32c91dbc56fd16e35886480ffdfb258ddc42a2fa