Do not write into SharedRef DB when git update fails
Data should not be written to the shared ref-db if the
git update did not succeed.

To allow a more manageable way to perform refs validation,
the logic and policy enforcement, previously distributed partially
in the MultiSiteBatchRefUpdate and ZkSharedRefDb,
is now centralised in a RefUpdateValidator class.

New metric counter introduced for every time a split-brain scenario
is detected and propagated as a more specific SharedDbSplitBrainException.

Bug: Issue 10711
Change-Id: I5395c22ffdf87fcc3247ea9a5d3afc57df60cfe9