Use pre-defined Gerrit EventGsonProvider
The current master of Gerrit heading to v3.1 has introduced auto-values
for many of the basic key objects, ranging from project keys, to user ids,
groups uuids and change ids.

Currently, only Change.Key is used in event classes and it is correctly
managed by the pre-defined EventGsonProvider. With regards to the other
keys, they are used only in cache evictions and are individually custom
deserialized because they are composed by a single field.

Moving forward, Gerrit should start introducing the automatic support
for AutoValue deserialization for all types; the custom de-serialization
of keys will then be removed and the EventGsonProvider will be used
across the board.

Change-Id: I279879daee8f12ce49ef6bc49f451d4f0726923f