Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2: Verified-1
> > > 
> > > I was actually wrong: it seems that the introduction of auto-values has broken the serializers :-(
> > 
> > @Dave any suggestion on how to deal with Gson de-serialization into auto-values? (example: ChangeEvnet has now changeKey and project fields as auto-values.
> 
> @Dave it seems the exception happens in the EventDeserializer: we could then have a genuine bug introduced inside Gerrit stream events.

Yes, confirmed: the same breakage happens on Gerrit *without* the multi-site plugin.
See the Change I242b51530a388 that demonstrates the breakage and the corresponding Issue 10776.

Let's wait for this problem to be fixed on master and, in the meantime, we'll have to disable the tests that are breaking for the same reason.

Patch-set: 2
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>