Update patch set 3
Patch Set 3: Verified+1

It's now working, by skipping the parts that are failing not because of this plugin but because of the Gerrit master Issue 10776.

Once the issue on Gerrit master is fixed, we can re-enable the tests.

Patch-set: 3
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=+1