Add project custom enforcement policy
Adds the ability of project or ref fine-grained policy
enforcement. This is typically needed whenever some of the projects
and refs requires different level of "consistency" across sites.

For example, the draft comments in All-Users are changed very often
across nodes and thus very likely to be misaligned. The loss of a draft
comment isn't nice, but is not so severe as the loss of a committed
code change.

Bug: Issue 10711
Change-Id: Ib477439b9c90af6e5c1719854e778c4b0aa6855e