Update patch set 4
Patch Set 4:

(4 comments)

Is this WIP? I believe we need a bit of extra description of what is the "shared-dir" and the "broker" in the overall architecture description.

Additionally, we would like to explain the rationale behind it and what are the benefits that it brings.

Last but not list, explain what is going to be a libModule and the default implementation and how the implementation of the broker and shared-dir are going to be plugged in.

Patch-set: 4
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>