MultiSiteRepository: Suppress deprecation warnings from overridden Repository methods
These methods are deprecated in favor of methods in RefDatabase. We
can't simply delegate to those because they return different types
and throw checked exceptions.

The intention in MultiSiteRepository is to wrap the Repository and
delegate all calls to the wrapped instance, so it's OK to implement
these deprecated methods. Just suppress the warnings.

Change-Id: I9ca954ff9bdeb441672cc1c9a46bdd6e03577909