Update patch set 5
Patch Set 5: Code-Review-1

(18 comments)

Looks good, however, there is a bit of confusion on who is preventing the split-brain. The Global Ref-DB is only responsible for telling the multi-site if a ref is up-to-date or not and for the compare-and-put atomically executed. Everything else is the responsibility of the multi-site libModule.

Adding Dave Borowitz to review it also, as today is its last day in the Gerrit Team :-)

P.S. Thanks again Dave for the useful discussion about multi-site last week in Germany.

Patch-set: 5
CC: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1