Introduce a NoOp Shared-Dir implementation
As part of the effort to extend the types of Share-Dir
implementations, the SharedRefDatabaseNoOp will perform no operations
against the Shared Reference Database.
The NoOp will be default interface binding in the Multi-Site library
and subsequently other plugins will override this implementation with
Zookeeper or other types of implementations.

Feature: Issue 10830
Change-Id: I694cc27a7954ee4d443c3e98a0e67969f542c66f