Decouple the multi-site engine from transport implementation
Use DynamicItem and DynamicSet with proper abstraction to decouple
multi-site engine from Kafka transport implementation. The multi-site
library would ship and bind a NoOp implementation, so that the whole
functionality is preserved.

Inside the multi-site.jar there is a Kafka implementation embedded
which can be loaded by linking the multi-site.jar from the
$GERRIT_SITE/lib into the $GERRIT_SITE/plugins and loaded as a plugin.

Feature: Issue 10829
Change-Id: I5f6e92e1a5470b4699d73cc19f5d788a985e8bd0