Make sure to always compare the latest local ref
When validating the ref operations against the shared ref-db,
make sure that values are always compared to the latest ones on
the repository and not the ones read and cached at the very beginning
of the operation.

Pushes can be transparently retried multiple times when there are
race conditions, the value read at the very beginning could not be
actual anymore when the ref-update is actually executed.

Change-Id: I7ac82f6f5437f60f3998d0ab3807434dcf95279a