Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review-1
> 
> (2 comments)
> 
> I believe this is just the initial change of a series where *all* the Kafka classes are provided with a NoOps default implementation, correct?
> 
> Can you start drafting the others also and put them in the same topic?

I'm not sure that this is right moment to implement this feature. Problem is that we have to bind NoOp and then replace it with the Kafka implementation in the same module. This is causing issue with binding because Kafka module is created by guice before main module where we have itemOf and NoOp binding. I suggest to do it together with moving out Kafka module to separate plugin if we want to implement it now we have to refactor the way how we are installing modules

Patch-set: 4
Reviewer: Gerrit User 1083454 <1083454@173816e5-2b9a-37c3-8a2e-48639d4f1153>