Introduce sharedRef database log file
Log to file all mutations to the sharedRef database so that each
interaction (creation, update and deletion) is recorded for analysis,
triage and debugging of issues.

Note:
Because multi-site gets installed as libModule, rather than a plugin,
PluginLogFile interface cannot be used, thus log appenders are defined
in a new LibModuleLogFile.

Feature: Issue 11166
Change-Id: I940c48238d8c6b88a439bc8400692de87ece3400