Log blob updates in sharedref_log
Not every ref update is a commit, some refs in gerrit are actually
pointers to blob objects instead. This is the case for
refs/sequences/changes for example.

Before this change, updates of blob objects were not recorded in the
sharedref_log file: now they are recorded as BLOB updates.

Bug: Issue 11243
Change-Id: I1661489934a46a5160378bddae30a8929bbeb788