Log blob updates in sharedref_log
Not every ref update is a commit, some refs in gerrit are actually
pointers to blob objects instead. This is the case for
refs/sequences/changes for example.

Before this change, updates of blob objects were not recorded in the
sharedref_log file: now they are recorded as UPDATE_REF.
Note that, because these are not commits, the committer and commit
message will not be present in the log. For example:

{
    "ref_name":"refs/sequences/changes",
    "old_id":"56a6051ca2b02b04ef92d5150c9ef600403cb1de",
    "new_id":"b5045cc4046dbc1d7cafa4c603fd3cdf35dc5dde",
    "project_name":"All-Projects",
    "type":"UPDATE_REF"
}

Bug: Issue 11243
Change-Id: I1661489934a46a5160378bddae30a8929bbeb788