Log acquiring and releasing sharedref locks
Before mutating refs entries in the sharedref database, a lock is
acquired to manage concurrency and guarantee exclusive access to that
ref. The lock is then released once the ref
has been created/updated/deleted from the sharedref database.

Add logging of lock acquisition/release to the sharedref_log, since
locks are also held in the sharedref database.

Feature: Issue 11254
Change-Id: Ie9f0d0d3422d5a6c8191703b472db4aaadfba974