Update patch set 7
Patch Set 7:

I tried to look into this to get it implemented by using DynamicItem, but unfortunately, I had no success.

DynamicItems are designed to be modified as plugins reload [1], which is only done in one place by the PluginGuiceEnvironment[2], on start and reload of plugins.

I am not sure doing a hot-swapping _within_ the same plugin is possible at all.
I would require to pass around ReloadableRegistrationhandle<> (which *cannot* be obtained by bind()) and, even if can be done, overall would result in a convoluted solution that would need to be disposed of soon after.

I suspect the separation of the zookeper implementation into a separate plugin might be a requirement to get this done.

Perhaps I am missing something and I just don't understand this well enough yet, but I would like your opinion on this.

@Luca, @Marcin, @Tiago thoughts?

[1] https://www.mvndoc.com/c/com.google.gerrit/gerrit-extension-api/com/google/gerrit/extensions/registration/DynamicItem.html

[2] https://github.com/atdt/gerrit/blob/master/gerrit-server/src/main/java/com/google/gerrit/server/plugins/PluginGuiceEnvironment.java#L351

Patch-set: 7
Reviewer: Gerrit User 1072905 <1072905@173816e5-2b9a-37c3-8a2e-48639d4f1153>