Update patch set 7
Patch Set 7: Code-Review-1

> Patch Set 7:
> 
> I tried to look into this to get it implemented by using DynamicItem, but unfortunately, I had no success.
> 
> DynamicItems are designed to be modified as plugins reload [1],

Or by other plugins during load.

> which is only done in one place by the PluginGuiceEnvironment[2], on start and reload of plugins.

Or the initial load also.

> I am not sure doing a hot-swapping _within_ the same plugin is possible at all.

What's the point of doing in the same plugin? Agree.
The object is to define a dynamic item in the multi-site libModule and then implement this in the zookeeper (or etc or other) plugins.

> I would require to pass around ReloadableRegistrationhandle<> (which *cannot* be obtained by bind()) and, even if can be done, overall would result in a convoluted solution that would need to be disposed of soon after.
> 
> I suspect the separation of the zookeper implementation into a separate plugin might be a requirement to get this done.

Or, just check if DynamicItem returns null and then do nothing.
A lot easier.

Patch-set: 7
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1