Bind SharedRefDB as DynamicItem to allow other implementations
Decouple the multi-site engine from the shared ref-db implementation
by using a DynamicItem. The multi-site library would ship and bind
a NoOp impleentation, so that the whole functionality is preserved.

Inside the multi-site.jar there is a Zookeeper implementation embendded
which can be loaded by linking the multi-site.jar from the $GERRIT_SITE/lib
into the $GERRIT_SITE/plugins and loaded as a plugin.

Change-Id: I7f978b9ba4edf81bf5a170e2eb9a5829edabef2e