Inject SharedRefDatabaseWrapper at startup
Since stable-3.0, validation module is installed in dbInjector
while the DynamicItem<SharedRefDatabase> is in sysInjector.
This would make DynamicItem.bind() tricky because should be defined
in sysInjector which isn't related to dbInjector.

Workaround the problem by defining optional the injector of
the DynamicItem<SharedRefDatabase> and delegate a new PluginStartup
lifecycle listener to complete the injectors once both dbInjector
and sysInjector are fully loaded.

Change-Id: I2127224ba1700beda9df36002fd08220948b0c27