Decouple the multi-site engine from borker implementation
Use DynamicItem and DynamicSet with proper abstraction to decouple
multi-site engine from Kafka broker implementation. The multi-site
library would ship and bind a NoOp implementation, so that the whole
functionality is preserved.

Inside the multi-site.jar there is a Kafka implementation embedded
which can be loaded by linking the multi-site.jar from the
$GERRIT_SITE/lib into the $GERRIT_SITE/plugins and loaded as a plugin.

Feature: Issue 10829
Change-Id: Ib05beb2c52bf5012f017c3bc114b7a892a1b648d