Decouple the multi-site engine from broker implementation
Use DynamicItem and DynamicSet with proper abstraction to decouple
multi-site engine from Kafka broker implementation.
Buind a NoOp implementation of the broker in The multi-site libModule
so that can be replaced by any plugin later on.

Inside the multi-site.jar there is a Kafka implementation of the broker
that can be loaded as a plugin.

Feature: Issue 10829
Change-Id: Ib05beb2c52bf5012f017c3bc114b7a892a1b648d