Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> All those configuration settings are redundant as I do not see the point of disabling the management of certain cache invalidations or reindexing.
> 
> Also, for the publisher and subscriber, I do not see the point of having them disabled either.
> 
> Can we just get rid of the configuration settings altogether by saying in the documentation "publisher and subscribers cannot be disabled independently anymore" and the same for the listeners?

Wow, that's a change in the design :) This will simplify the code so it's a good choice. One question do we want to have one common general setting to disable publisher and subscriber or we don't want to allow that?

Patch-set: 1
Reviewer: Gerrit User 1083454 <1083454@173816e5-2b9a-37c3-8a2e-48639d4f1153>