Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > I'm not sure that we want to change formatting. But if we agree that we should we have to do it on stable-2.16(or stable-3.0 if only fixes are allowed on stable-2.16)
> 
> Can we do it in master and then apply it to the other branches?
> 
> Tab formatting is forbidden nowadays 😄

I understand, but then merging fixes from stable-2.16 will be more complicated but I'm fine with doing that. Our rule of thumb is that all development is done as earl as possible and then merge up. This allows to have clean history and make sure that we are not missing anything, so I advise :) to do that on stable-3.0

Patch-set: 1
Reviewer: Gerrit User 1083454 <1083454@173816e5-2b9a-37c3-8a2e-48639d4f1153>