Index account asynchronously and flush upon replication
Account reindexing on remote nodes should not happen immediately
but must queue for asynchronous execution once replication is completed.

If replication has failed, do not flush the indexing queue.

TODO: there is not yet a reliable way to understand IF the replication
target node is the one receiving the event. Trigger reindexing anyway
because it is less of a problem to reindex an account twice than missing
a reindexing event.

Bug: Issue 12278
Change-Id: Ie442b03f36c1238f6b977e7f5456beff078c9657