Index account asynchronously and flush upon replication
Account reindexing on remote nodes should not happen immediately
but must queue for asynchronous execution once replication is completed.

Triggering the reindex of the account *before* it arrives on the
All-Users repo would generate two issues:

1. The account would be reindexed with stale data and, when the new
   data will arrive, it won't be reindexed again: we do need to wait
   for the data to arrive through replication.

2. The account index would also populate the cache with stale data
   and even when the new data will arrive the in-memory data would need
   to wait the cache TTL before being visible to the end-user.

NOTE: there is not yet a reliable way to understand IF the replication
target node is the one receiving the event. Trigger reindexing anyway
because it is less of a problem to reindex an account twice than missing
a reindexing event.

Bug: Issue 12278
Change-Id: Ie442b03f36c1238f6b977e7f5456beff078c9657