Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What's the benefit of using file-based replication over SSH?

Works out of the box, allowing people to experiment the plugin locally without prior configuration.

> Is there any gotcha implied by the SSH replication that creates problems?

Requires configuration:
- defining keys in a JSch-compatible format
- defining .ssh/config to ignore host key checking
- removing in local .ssh/config all options not supported by JSch

> I was under the impression SSH replication was a good choice because it was going to be more adherent to an actual installation of the replication plugin

For the purpose of testing the multi-site plugin, the protocol of replication is 100% irrelevant: I would like to make the people's life easier when using this plugin in a dev environment.

Patch-set: 1