Fix NPE when the replicated refs is not found locally
When the replication done events are received, the local ref
may or may not be replicated locally.
Do not fail with NPE but just trace a warning on the error log
saying what happened.

Also, do not act on ref-replicated event but rather on the
ref-replication-done event, which gives more guarantee that
the replication has successfully reached all nodes.

Change-Id: I50acc663b586cf2191a74b577ead0ffdbfb6f363