Update project version on global ref-db only when more recent
The project version (refs/multi-site/version) on the local project is more
up-to-date than the one on global ref-db, update unconditionally
without requiring the local SHA1 to be the same of the global ref-db.

When refs are updated in a short period of time (e.g. refs/multi-site/version
and refs/changes/N/NNNN/meta) it is not guaranteed that the replication
will be able to push all the intermediate SHA1s. Skipping SHA1s to push
for non-fast-forward branches like refs/multi-site/version is a valid
use-case and thus can be performed as long as the local version is
the most recent one.

Failing to update the global ref-db project version causes split-brain
with all the other nodes because the replication would filter out
the refs out of sync.

Bug: Issue 12413
Change-Id: Ib5e49edd3ba37ed60f34736f7af5e5a324248e51