Use a global version number instead of the commit timestamp
The commit timestamp is not a reliable source of repository
version numbering because of unreliability of the local
time of the committer and forced push use-case, where the
repository is brought back in time.

Define a new global version numbering
(current system time at the moment) and
use it as monotonic versioning for the repository.

Bug: Issue 12413
Change-Id: If1b642a56791bd3a4c1d6ef903ec2f7c61c19928