Reload local version when in memory ref is not up to date
When multiple updates are replication events might pile up and the ref
carried in the RemoteRefUpdate event might be behind the one locally
stored.

This will lead to filter out some replication events. The situation is
restored as soon as a new replication will be triggered.
The repository will be out of sync until the arrival of the next
replication.

Read the local ref and re-try the comparison to mitigate the issue.

Bug: Issue 12425
Change-Id: I579f277635eff5d5cdd1d71a090fe06f2be22c0f