DeleteChangeUsingMultiGerrit1: Fix default port typo
The default port for this scenario should have been 8081 originally.

Testing the CreateChangeUsingMultiGerrit scenario locally using the
dockerised_local_env, which uses this DeleteChangeUsingMultiGerrit1
scenario, doesn't lend a full green report. The latter scenario fails
with 401 unauthorized, with this commit; using the previous 8082 port
passed it all. However 8081 (this commit) *is* the required test port.

Hence leave the below Issue open for now or until that 401 issue with
the dockerised_local_env is resolved. Trying to re-generate the test
user (admin's) HTTP password or restarting the composed SUT containers
didn't help, locally.

Bug: Issue 12693
Change-Id: Ib67d24c26d6148bfb8caca2b6d6c00ec51a60911