Do not mock Cache<?,?>
Cache is annotated with DoNotMock and therefore should not
be mocked in tests but rather created from a builder and called
with assertions.

It was previously mocked until now but, since Gerrit v3.2,
a more restrictive ErrorProne policy checks for non-mockable
interfaces and reports an error for the Cache mocks.

Change-Id: Ic6de322bf5f18e567a174ee88a05696f2db2dcff