Set forward context for events not being produced locally
The multi-site plugin has a concept of forwarded event.
This is modelled as a Context property and it is checked on
the consuming side of the plugin, so that consumed events are
flagged as forwarded and not re-forwarded.

Set the forward context at the _producing_ side as well, so that events
that were received through different channels (i.e. high-availability)
are not re-forwarded.

The event source check is done by checking whether the event payload
carries the instanceId of a gerrit instance different from the current
one.

Bug: Issue 13308
Change-Id: I9dc270a1cc29647844e45becefa20389ce561678