Populate MultiSite events' instanceId
Since stable-3.2 Gerrit introduces the new instanceId configuration
which allows to identify a specific instance within a group of Gerrit
instances.

When such instanceId is set, Gerrit makes sure to populate the Event
field `instanceId` before distributing the events to relevant listeners.

When this happens, however, the event has already been handed over by
the multi-site broker, so it is not possible for the multi-site plugin
to utilize the instanceId at all.

Ensure to populate MultiSite events' instanceId *as soon as* they get
created, and thus *before* they get dispatched, so that multi-site can
apply arbitrary logic, based on that value.

Bug: Issue 13308
Change-Id: I41101648664bd23b029d67fec3591c776c2b7de8