Populate MultiSite events' instanceId
Since stable-3.2 Gerrit introduces the new instanceId configuration
which allows to identify a specific instance within a group of Gerrit
instances.

When such instanceId is set, Gerrit makes sure to populate the Event
field `instanceId` before distributing the events to relevant listeners.

Populate the `instanceId` for all the other non-Gerrit events,
such as cache evictions, reindexing and project creation/deletion,
so that all events can consistently have the source instanceId
for preventing events loops.

Gerrit instanceId is optional, so when it is omitted, use the multi-site
UUID instanceId as a surrogate.

Change-Id: I5845597cb2d057f28f031f72984da2f7ea233d7c