Do not forward events from high-availability
Do not forward to other sites the events generated locally
but as a result of the execution of an high-availability API.

All the Gerrit instances are connected to multi-site, so events
forwarded by the high-availability plugin do not require any
further propagation to other sites.

Only the events *directly* generated by the local site and
not forwarded by the high-availability will continue to be
forwarded as usual.

Bug: Issue 13307
Change-Id: I8e41d5aaf3f61c03a491e06022b50a2cc08995b7