Do not forward events from high-availability
Do not forward to other sites The events generated locally
but as a result of the execution of an high-availability API.

All the Gerrit instances are connected to multi-site, so events
forwarded by the high-availability plugin do not require any
further propagation to other sites.

Only the events *directly* generated by the local site and
not forwarded by the high-availability will continue to be
forwarded as usual.

Bug: Issue 13307
Change-Id: I9093c055272648d63045f43d5e26863883de8089