Ensure All-Users:refs/meta/external-ids doesn't cause split-brains
The DESIRED policy was introduced in the early stages of the multi-site
to work around locking problems with the global ref-database, which have
later been fixed with the introduction of a locking mechanism[1].

Due to the concurrent nature of All-Users:refs/meta/external-ids, the
DESIRED policy had been applied by default to avoid causing continuous
client failures.

The entire concept behind DESIRED policy however, is wrong and prone to
cause spit-brains due to the fact that nothing, really, is done to
prevent them: this should not be applied to a crucial core ref such as
refs/meta/external-ids.

Remove DESIRED policy (that was only used for
All-Users:refs/meta/external-ids) and instead make sure that it is
REQUIRED to be in sync with the global ref-database.

[1] I5395c22ffdf87fcc3247ea9a5d3afc57df60cfe9

Issue: Bug 13462
Change-Id: I6c8fb571e5467fd5596fb8ee1fcdf0179350a051