Stop broadcasting accounts cache eviction
Since the introduction of CachedAccountDetails[1], the sha1 of the
refs/users/<shared-id> is used as cache key for the accounts cache.

An effect of this is that explicit cache invalidations are not required
anymore, since changing account details generates a new sha1,
effectively invalidating the previous entry.

Accounts cache invalidation could still be achieved by explicitly
flushing caches, but the effects of this operation should not be
propagated to all nodes, since this would make all nodes cold,
potentially affecting gerrit performance across all sites.

Stop accounts cache evictions from being propagated to the broker.

[1] https://gerrit-review.googlesource.com/c/gerrit/+/260992

Bug: Issue 13820
Change-Id: I9cd1c95693c9d542ffe2871637619068eeed233c