Stop broadcasting accounts cache eviction
Since the introduction of CachedAccountDetails (I61ae5780), the sha1 of
the refs/users/<shared-id> is used as part of the cache key for the
accounts cache.

An effect of this is that explicit cache invalidations are not required
anymore, since changing account details generates a new sha1,
effectively invalidating the previous entry.

Accounts cache invalidation could still be achieved by explicitly
flushing caches, but the effects of this operation should not be
propagated to all nodes, since this would make all nodes cold,
potentially affecting gerrit performance across all sites.

Stop accounts cache evictions from being propagated to the broker.

Bug: Issue 13820
Change-Id: I9cd1c95693c9d542ffe2871637619068eeed233c