Fix injection of the ignoredRef for global-refdb validation
The ignoredRef must be injected as ImmutableSet<String> rather
than Set<String>. The change to the ImmutableSet happened during
the last review of the global-refdb, but the corresponding change
in multi-site was left behind.

This has passed unnoticed because the ignoredRef injection is
optional and therefore just continued with the default value
which is empty = do not ignore any refs.

Change-Id: Ib8f756dc4b336af25149fa4bb6fd5a57acbe1fdb