Test cache invalidation behaviour inside a Gerrit daemon
All existing tests for cache forwarding were not plugging
the real Gerrit caches. A change happened in the project cache
in stable-3.3 broke the cache invalidation without breaking
the tests.

Perform an effective E2E cache invalidation test from
a plugin integration test environment, so that if any of
the cache keys are changing, the test would allow to
identify the breakage and the associated cache involved.

The first cache included in the E2E test is 'projects'.
Other caches are going to be covered in the follow-up
changes.

Bug: Issue 14255
Change-Id: I94db246f6310f898ee673dc1a9bad31ffedeef83