Allow kinesis-events to be configured as events-broker
Originally kafka was the only implementation of events-broker, so it
made sense that the local setup of a multi-site environment was
hardcoding kafka as the only events-broker implementation.

With the upcoming of additional implementations however, the user should
be able to specify which flavour of broker implementation should be
installed.

Introduce the ability to install kinesis as opposed to kafka.

Bug: Issue 14274
Change-Id: I9c6be9ffeaca02116cd17759d5951581f3585336