Wait for groups replication before attempting indexing
Do not immediately attempt to (re)index a group before the related has
landed in git, which might take some time depending on replication
delays and settings.

Similarly to the ForwardedIndexChangeHandler, the
ForwardedGroupChangeHandler will attempt multiple reindexes up to
`index.maxTries`, every `index.retryInterval` milliseconds.

To be able to understand whether replication is completed and the local
group ref is up-to-date, two main changes have been introduced:

- A GroupChecker class, with the responsibility to lookup in All-Users
  for the status of a specified group
- An extra field in the GroupIndexEvent class, to indicate which
  particular sha1 the ref group was updated to, so that GroupChecker
  could perform a specific lookup.

Note that to maintain back-compatibility with previous
GroupIndexEvent's, the additional field was made optional, so that old
events can still be de-serialized and consumed by this version of the
multi-site code.

Bug: Issue 14161
Change-Id: Ifc171a49a0c360ceb39538a6d133154fcd0050e9